+++ R/utils-vignettes.R
@@ -1,7 +1,7 @@
 # Wrappers to use in vignette building for R 3.0.0
 
 vweave = vtangle = function(file, driver, syntax, encoding = '', quiet = FALSE, ...) {
-  opts_knit$set(stop_on_error = 2L)  # should not hide errors
+  opts_chunk$set(error = FALSE)  # should not hide errors
   options(markdown.HTML.header = NULL)
   (if (grepl('\\.[Rr]md$', file)) knit2html else if (grepl('\\.[Rr]rst$', file)) knit2pdf else knit)(
     file, encoding = encoding, quiet = quiet, envir = globalenv()