+++ NEWS.md
@@ -14,6 +14,8 @@
 
 - **knitr** formally depends on the **highr** package now (for syntax highlighting of LaTeX and HTML)
 
+- the package option `stop_on_error` has been deprecated; now it is much easier to specify whether you want to stop on errors or not by using the existing chunk option `error`; if you want to stop, use `error=FALSE`; see the documentation for details: http://yihui.name/knitr/options
+
 ## MINOR CHANGES
 
 - for the chunk options set in package option `opts_knit$get('eval.after')`, they will not be evaluated after a chunk if `eval=FALSE` for that chunk (#570, thanks, @knokknok)